master -> web

+++ lmfdb/backend/statstable.py
@@ -278,6 +278,27 @@ class PostgresStatsTable(PostgresBase):
         if cur.rowcount:
             return int(cur.fetchone()[0])
 
+    def null_counts(self, suffix=""):
+        """
+        Returns the columns with null values, together with the count of the number of null rows for each
+        """
+        selecter = SQL(
+            "SELECT cols, count FROM {0} WHERE values = %s AND split = %s"
+        ).format(Identifier(self.counts + suffix))
+        cur = self._execute(selecter, [Json([None]), False])
+        allcounts = {rec[0][0]: rec[1] for rec in cur}
+        for col in self.table.search_cols:
+            if col not in allcounts:
+                allcounts[col] = self._slow_count({col: None}, suffix=suffix, extra=False)
+        return {col: cnt for col,cnt in allcounts.items() if cnt > 0}
+
+    def refresh_null_counts(self, suffix=""):
+        """
+        Recomputes the counts of null values for all search columns
+        """
+        for col in self.table.search_cols:
+            self._slow_count({col: None}, suffix=suffix, extra=False)
+
     def _slow_count(self, query, split_list=False, record=True, suffix="", extra=True):
         """
         No shortcuts: actually count the rows in the search table.
@@ -320,15 +341,17 @@ class PostgresStatsTable(PostgresBase):
             used to store the count
         - ``extra`` -- see the discussion at the top of this class.
         """
+        # We only want to record 0 counts for value [NULL], since other cases can break stats
+        nullrec = (list(query.values()) == [None])
         cols, vals = self._split_dict(query)
         data = [count, cols, vals, split_list]
         if self.quick_count(query, suffix=suffix) is None:
-            if count == 0:
+            if count == 0 and not nullrec:
                 return # we don't want to store 0 counts since it can break stats
             updater = SQL("INSERT INTO {0} (count, cols, values, split, extra) VALUES (%s, %s, %s, %s, %s)")
             data.append(extra)
         else:
-            if count == 0:
+            if count == 0 and not nullrec:
                 updater = SQL("DELETE FROM {0} WHERE cols = %s AND values = %s AND split = %s")
                 data = [cols, vals, split_list]
             else:
@@ -483,7 +506,7 @@ class PostgresStatsTable(PostgresBase):
             nres = self._slow_count_distinct(col, query, record=record)
         return int(nres)
 
-    def column_counts(self, cols, constraint=None, threshold=None, split_list=False):
+    def column_counts(self, cols, constraint=None, threshold=1, split_list=False):
         """
         Returns all of the counts for a given column or set of columns.
 
@@ -1461,9 +1484,13 @@ ORDER BY v.ord LIMIT %s"""
                                 curlevel.append((colvec + col, j))
                     level += 1
 
-    def _status(self):
+    def _status(self, reset_None_to_1=False):
         """
         Returns information that can be used to recreate the statistics table.
+
+        INPUT:
+
+        - ``reset_None_to_1`` -- change threshold None to 1 in the stored statistics
 
         OUTPUT:
 
@@ -1487,9 +1514,16 @@ ORDER BY v.ord LIMIT %s"""
                 grouping = cgcols[len(cvals) :]
                 ccols = cgcols[: len(cvals)]
             nstat_cmds.append((cols[0], grouping, ccols, cvals, threshold))
+        if reset_None_to_1:
+            for L in [stat_cmds, split_cmds, nstat_cmds]:
+                for i in range(len(L)):
+                    newval = list(L[i])
+                    if newval[-1] is None:
+                        newval[-1] = 1
+                        L[i] = tuple(newval)
         return stat_cmds, split_cmds, nstat_cmds
 
-    def refresh_stats(self, total=True, suffix=""):
+    def refresh_stats(self, total=True, reset_None_to_1=False, suffix=""):
         """
         Regenerate stats and counts, using rows with ``stat = "total"`` in the stats
         table to determine which stats to recompute, and the rows with ``extra = True``
@@ -1499,6 +1533,7 @@ ORDER BY v.ord LIMIT %s"""
 
         - ``total`` -- if False, doesn't update the total count (since we can often
             update the total cheaply)
+        - ``reset_None_to_1`` -- change threshold None to 1 in stored statistics
         - ``suffix`` -- appended to the table name when computing and storing stats.
             Used when reloading a table.
         """
@@ -1506,7 +1541,7 @@ ORDER BY v.ord LIMIT %s"""
         t0 = time.time()
         with DelayCommit(self, silence=True):
             # Determine the stats and counts currently recorded
-            stat_cmds, split_cmds, nstat_cmds = self._status()
+            stat_cmds, split_cmds, nstat_cmds = self._status(reset_None_to_1)
             col_value_dict = self.extra_counts(include_counts=False, suffix=suffix)
 
             # Delete all stats and counts
@@ -1528,11 +1563,11 @@ ORDER BY v.ord LIMIT %s"""
                 self.total = self._slow_count({}, suffix=suffix, extra=False)
             self.logger.info("Refreshed statistics in %.3f secs" % (time.time() - t0))
 
-    def status(self):
+    def status(self, reset_None_to_1=False):
         """
         Prints a status report on the statistics for this table.
         """
-        stat_cmds, split_cmds, nstat_cmds = self._status()
+        stat_cmds, split_cmds, nstat_cmds = self._status(reset_None_to_1)
         col_value_dict = self.extra_counts(include_counts=False)
         have_stats = stat_cmds or split_cmds or nstat_cmds
         if have_stats:
@@ -1667,7 +1702,7 @@ ORDER BY v.ord LIMIT %s"""
         - ``header`` -- a list of lists giving the values to print along the top or side of the table
         - ``data`` -- a dictionary with data on counts
         """
-        selecter_constraints = [SQL("split = %s"), SQL("cols = %s")]
+        selecter_constraints = [SQL("split = %s"), SQL("cols = %s"), SQL("count > 0")]
         if constraint:
             allcols = sorted(set(cols + list(constraint)))
             selecter_values = [split_list, Json(allcols)]
@@ -1697,8 +1732,6 @@ ORDER BY v.ord LIMIT %s"""
                     else:
                         selecter_constraints.append(SQL("values->{0} = %s".format(i)))
                         selecter_values.append(Json(cx))
-                    # count 0 rows shouldn't be added usually, but if they get in it can cause havoc since some formatters hard-code the valid inputs
-                    selecter_constraints.append(SQL("count > 0"))
         else:
             allcols = sorted(cols)
             selecter_values = [split_list, Json(allcols)]
